// src/app/models/cliente.ts
export interface Cliente {
    id?: any;
    cnpj: String;
    nome: String;
    cep: String;
    tipo: String;
    logradouro: String;
    numero: String;
    complemento: String;
    bairro: String;
    uf: String;
    municipio: String;
    email: String;
    telefoneResidencial: String;
    obs: String;
}